Add proof of concept get translation API endpoint
Add API endpoint that gets the translation of a name in a specified
language. This is a very rough proof of concept; it works, but it's
not DRY and probably has a lot of redundant and messy code. I don't
really understand JavaScript promises, so this was pieced together by
copy pasting the get name code and tweaking it to find the
translation.

This only supports ISO 639-3 codes, although I would like to expand it
to support full language names as well.

The output format is also a bit of a proof of concept. I'm not sure it
makes sense to return the name, although I think returning the full
translation object is useful.